refactor(claw-server-rust): move services into subsystems
