+++ packages/browseros-agent/apps/claw-server-rust/src/app.rs
@@ -1,16 +1,17 @@
 use crate::{
+    browser::BrowserService,
+    capture::{
+        audit::AuditService, recordings::RecordingStore,
+        replays::ReplayService as ReplayReadService, screencast::ScreencastService,
+        screenshots::ScreenshotService,
+    },
     config::Config,
     error::AppResult,
     routes,
-    services::{
-        agents::AgentService, audit::AuditService, browser::BrowserService,
-        harness::HarnessService, recordings::RecordingStore,
-        replays::ReplayService as ReplayReadService, screencast::ScreencastService,
-        screenshots::ScreenshotService, tab_activity::TabActivityService,
-        tab_targets::TabTargetMap, telemetry::TelemetryService,
-    },
+    services::{agents::AgentService, harness::HarnessService, telemetry::TelemetryService},
     sessions::Sessions,
     storage::JsonStore,
+    tabs::{activity::TabActivityService, targets::TabTargetMap},
 };
 use axum::{Router, middleware};
 use std::{env, path::PathBuf, sync::Arc, time::Duration};
