+++ packages/browseros-agent/apps/claw-server-rust/src/capture/mod.rs
@@ -1,0 +1,5 @@
+pub mod audit;
+pub mod recordings;
+pub mod replays;
+pub mod screencast;
+pub mod screenshots;
