+++ packages/browseros-agent/apps/claw-server-rust/src/capture/recordings.rs
@@ -1,10 +1,10 @@
 use crate::{
+    capture::audit::AuditService,
     db::audit::entities::{
         prelude::{TabClaims, TabRecordings},
         tab_claims, tab_recordings,
     },
     error::{AppError, AppResult, IoPath},
-    services::audit::AuditService,
 };
 use sea_orm::{
     ActiveValue::Set,
@@ -261,7 +261,7 @@ impl RecordingStore {
             loop {
                 ticker.tick().await;
                 match self
-                    .sweep_retention(retention_days, super::now_epoch_ms())
+                    .sweep_retention(retention_days, crate::services::now_epoch_ms())
                     .await
                 {
                     Ok(result) => info!(
@@ -602,11 +602,11 @@ fn sanitize_target_id(target_id: &str) -> String {
 mod tests {
     use super::{BATCH_ID_LRU_CAPACITY, RecordedEvent, RecordingEventInput, RecordingStore};
     use crate::{
+        capture::audit::AuditService,
         db::audit::entities::{
             prelude::{TabClaims, TabRecordings},
             tab_claims,
         },
-        services::audit::AuditService,
     };
     use sea_orm::{
         ActiveValue::{NotSet, Set},
