+++ packages/browseros-agent/apps/claw-server-rust/src/capture/screenshots.rs
