+++ packages/browseros-agent/apps/claw-server-rust/src/mcp/effects/ownership_claims.rs
@@ -25,7 +25,7 @@ pub fn apply(
                     .call
                     .state
                     .tab_activity
-                    .record_tool(crate::services::tab_activity::RecordToolInput {
+                    .record_tool(crate::tabs::activity::RecordToolInput {
                         target_id: info.target_id,
                         page_id,
                         url: info.url,
