+++ packages/browseros-agent/apps/claw-server-rust/src/tabs/activity.rs
